Hoist domain-agnostic infra into src/framework/
Step 1 of the structural reorg: collapse the three agnostic top-level
directories (src/core/, src/middleware/) and the generic SQLAlchemy
infra under src/models/ (base.py, audit_log.py, _polymorphic.py) into
src/framework/. The framework now owns every domain-agnostic primitive.

Moves:
  src/core/config.py            -> src/framework/config.py
  src/core/templating.py        -> src/framework/templating.py
  src/core/form_fields.py       -> src/framework/form_fields.py
  src/core/test_*.py            -> src/framework/test_*.py
  src/models/base.py            -> src/framework/base_model.py
  src/models/audit_log.py       -> src/framework/audit_log.py
  src/models/_polymorphic.py    -> src/framework/polymorphic.py
  src/models/test__polymorphic.py -> src/framework/test_polymorphic.py

Deletes:
  src/middleware/  (empty placeholder; real middleware in framework/)
  src/core/        (now empty)

src/models/__init__.py keeps re-exporting Base / BaseModel / metadata /
AuditLog from their new framework home so `from src.models import X`
keeps working — the hub stays stable while the implementation moves.

Broke a latent circular import (framework -> models -> framework) by
moving the `from src.models import User` lines in framework/authz.py
and framework/responses.py under TYPE_CHECKING with
`from __future__ import annotations`. User is only used as a type hint
in those files.
