+++ config/README.md
@@ -1,3 +1,3 @@
 # Config
 
-`alembic.ini` lives here (Alembic's own convention). Application configuration is in [`src/core/`](../src/core/README.md); migrations themselves are in [`alembic/`](../alembic/README.md).
+`alembic.ini` lives here (Alembic's own convention). Application configuration is in [`src/framework/config.py`](../src/framework/config.py); migrations themselves are in [`alembic/`](../alembic/README.md).
