--- src/middleware/README.md
@@ -1,3 +1,0 @@
-# Middleware
-
-Reserved for future ASGI middleware. Currently active middleware lives in [`src/framework/middleware.py`](../framework/middleware.py) and is registered in [`src/main.py`](../main.py).
