+++ src/models/providers/provider_education.py
@@ -3,7 +3,8 @@ from functools import partial
 from sqlalchemy import Column, ForeignKey, Text
 from sqlalchemy.types import Uuid
 
-from ..base import BaseModel
+from src.framework.base_model import BaseModel
+
 from ..enums import EDUCATION_TYPES, named_check_in
 
 _TABLE = "provider_educations"
