+++ src/models/providers/provider_licensure.py
@@ -3,7 +3,8 @@ from functools import partial
 from sqlalchemy import Column, Date, ForeignKey, Text
 from sqlalchemy.types import Uuid
 
-from ..base import BaseModel
+from src.framework.base_model import BaseModel
+
 from ..enums import LICENSE_TYPES, US_STATES, named_check_in
 
 _TABLE = "provider_licensures"
