+++ src/models/users/user.py
@@ -3,7 +3,7 @@ import uuid
 from fastapi_users.db import SQLAlchemyBaseUserTable
 from sqlalchemy import Column, Text
 
-from ..base import BaseModel
+from src.framework.base_model import BaseModel
 
 
 class User(SQLAlchemyBaseUserTable[uuid.UUID], BaseModel):
