+++ src/templates/posts/README.md
@@ -22,4 +22,4 @@ Templates dereference through the right relationship per branch: `{{ d.descripti
 1. Add the per-kind form-body partial `posts/_<kind>_form.html` (a `{% macro %}` taking `(hx_method, action, submit_label, post=None)`) using the shared field-render macros in [`../_shared/form_fields.html`](../_shared/form_fields.html).
 2. Add `posts/new_<kind>.html` and `posts/edit_<kind>.html` as ~5-line wrappers around the form macro. Register their paths on the kind's `PostKindSpec` in [`src/models/posts/post_kinds.py`](../../models/posts/post_kinds.py); the route layer reads `spec.create_template` / `spec.edit_template` from there.
 3. Add a `{% elif post.kind == "<kind>" %}` branch in `list.html` and `detail.html` for the per-kind row label and read-view body. (The "New X" link in `list.html` follows automatically via the registry-driven loop.)
-4. If the kind introduces a new controlled-vocabulary, add the tuple + `*_LABELS` dict to `src/models/enums.py` and register both as Jinja globals in `src/core/templating.py`. The shared `select_field` macro then renders dropdowns from them with no further per-template work.
+4. If the kind introduces a new controlled-vocabulary, add the tuple + `*_LABELS` dict to `src/models/enums.py` and register both as Jinja globals in `src/framework/templating.py`. The shared `select_field` macro then renders dropdowns from them with no further per-template work.
